feat: :key: Configure API automation user and permissions
This commit introduces the configuration for an API automation user (packer@pve) and the corresponding group (api-automation) within Proxmox. It also sets up the necessary ACLs to grant the group appropriate permissions for VM administration and storage access. This allows for automated VM creation and management via tools like Packer and Terraform.